Add finish clauses where applicable to clean up app control flow, also file organization refactor and add hook to tutorial inside settings activity
