+++ app/src/main/java/com/cs407/beet_boxing/activities/HelpActivity.java
@@ -1,10 +1,12 @@
-package com.cs407.beet_boxing;
+package com.cs407.beet_boxing.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.content.Intent;
 import android.os.Bundle;
 import android.widget.Button;
+
+import com.cs407.beet_boxing.R;
 
 public class HelpActivity extends AppCompatActivity {
 
@@ -14,6 +16,9 @@ public class HelpActivity extends AppCompatActivity {
         setContentView(R.layout.activity_help);
 
         Button returnButton = findViewById(R.id.returnButton);
-        returnButton.setOnClickListener(e -> startActivity(new Intent(this, MainActivity.class)));
+        returnButton.setOnClickListener(e -> {
+            startActivity(new Intent(this, MainActivity.class));
+            finish();
+        });
     }
 }
