+++ app/src/main/res/layout/activity_garden.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ActivityGarden">
+    tools:context=".activities.GardenActivity">
 
     <include layout="@layout/edit_menu"/>
