+++ app/src/main/res/layout/activity_help.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HelpActivity">
+    tools:context=".activities.HelpActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
