+++ app/src/main/res/layout/activity_recording_mode.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RecordingModeActivity">
+    tools:context=".activities.RecordingModeActivity">
 
     <include
         android:id="@+id/include"
