+++ app/src/main/res/layout/activity_recordings_list.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RecordingsListActivity">
+    tools:context=".activities.RecordingsListActivity">
 
     <!-- Toolbar for the Activity -->
     <androidx.appcompat.widget.Toolbar
