+++ app/src/main/res/layout/activity_share_recording.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".share_recording">
+    tools:context=".activities.ShareRecordingActivity">
 
     <Button
         android:id="@+id/playButton"
