+++ app/src/main/res/layout/settings_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SettingsActivity">
+    tools:context=".activities.SettingsActivity">
 
     <TextView
         android:id="@+id/options"
@@ -55,7 +55,7 @@
         android:id="@+id/seekBarVolume"
         android:layout_width="233dp"
         android:layout_height="21dp"
-        android:layout_above="@+id/replayButton"
+        android:layout_above="@+id/tutorialButton"
         android:layout_below="@+id/volume"
         android:layout_alignParentStart="true"
         android:layout_alignParentEnd="true"
@@ -75,13 +75,13 @@
 
 
     <Button
-        android:id="@+id/replayButton"
+        android:id="@+id/tutorialButton"
         android:layout_width="197dp"
         android:layout_height="60dp"
         android:layout_alignParentBottom="true"
         android:layout_centerHorizontal="true"
         android:layout_marginBottom="126dp"
-        android:text="Replay Tutorial" />
+        android:text="Tutorial" />
 
     <Button
         android:id="@+id/backButton"
